Reworked REGEN detection
Nightly builds of BlastEM were being incorrectly read as REGEN because of two things:

1. Nightly versions of BlastEM patch the previous method of detection
2. The previous method of detection for REGEN is present on a variety of emulators, including BlastEM

As such, a more futureproof way of detecting REGEN was necessary. The new method relies on a bug related to YM Timers in REGEN that will almost surely never be patched out in an emulator that hasn't had a new release in 11 years. If it ever does, a mouse input fix had better be included with that or I will flip my lid.